+++ Code/Scripts/Mission04.cpp
@@ -3105,7 +3105,7 @@ DECLARE_SCRIPT(M04_EngineRoom_Prisoner_01_JDG, "")//this guys ID number is M04_P
 		{
 			if (damager == STAR )
 			{
-				int myMaxHealth = Commands->Get_Max_Health ( obj );
+				float myMaxHealth = Commands->Get_Max_Health ( obj );
 				Commands->Set_Health ( obj, myMaxHealth );
 
 				if (firstTimeDamaged == true)
@@ -3270,7 +3270,7 @@ DECLARE_SCRIPT(M04_EngineRoom_Prisoner_02_JDG, "")//this guys ID number is M04_P
 	{
 		if (obj && damager == STAR)
 		{
-			int myMaxHealth = Commands->Get_Max_Health ( obj );
+			float myMaxHealth = Commands->Get_Max_Health ( obj );
 			Commands->Set_Health ( obj, myMaxHealth );
 
 			if (firstTimeDamaged == true)
@@ -3378,7 +3378,7 @@ DECLARE_SCRIPT(M04_EngineRoom_Prisoner_03_JDG, "")//this guys ID number is M04_P
 		ActionParamsStruct params;
 		if (obj && damager == STAR)
 		{
-			int myMaxHealth = Commands->Get_Max_Health ( obj );
+			float myMaxHealth = Commands->Get_Max_Health ( obj );
 			Commands->Set_Health ( obj, myMaxHealth );
 
 			if (firstTimeDamaged == true)
